[#775] Rework closing logic of Worker/Boss
Before we just shutdown the Executor's which were used to serve the IO, now we only do so if we created them by our own. This means if you create a *ChannelFactory and pass in your own Executor, you will need to shutdown them by yourself after calling releaseExternalResources() on the ChannelFactory. The same is true when create WorkerPool and BossPool instances.

This commit also fixes the problem that the Worker/Boss Thread will never be released if you use an Executor (no ExecutorService).

+++ src/main/java/org/jboss/netty/channel/socket/oio/OioClientSocketChannelFactory.java
@@ -76,6 +76,7 @@ public class OioClientSocketChannelFactory implements ClientSocketChannelFactory
 
     private final Executor workerExecutor;
     final OioClientSocketPipelineSink sink;
+    private boolean shutdownExecutor;
 
     /**
      * Creates a new instance with a {@link Executors#newCachedThreadPool()} as worker executor.
@@ -84,6 +85,7 @@ public class OioClientSocketChannelFactory implements ClientSocketChannelFactory
      */
     public OioClientSocketChannelFactory() {
         this(Executors.newCachedThreadPool());
+        shutdownExecutor = true;
     }
 
     /**
@@ -105,6 +107,8 @@ public class OioClientSocketChannelFactory implements ClientSocketChannelFactory
     }
 
     public void releaseExternalResources() {
-        ExecutorUtil.terminate(workerExecutor);
+        if (shutdownExecutor) {
+            ExecutorUtil.terminate(workerExecutor);
+        }
     }
 }
